+++ libraries/template.rb
@@ -1,8 +1,6 @@
 #
 # Cookbook:: unbound
-# Resource:: rr
-#
-# Copyright:: 2011, Joshua Timberman
+# Library:: template
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -16,11 +14,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-property :fqdn,     String, name_property: true
-property :ip,       String, required: true
-property :type,     String, default: 'host'
-property :cwd,      String
-property :exists,   NilClass, default: nil
 
-action :add do
+module Unbound
+  module Cookbook
+    module TemplateHelpers
+      def template_partial_indent(output, level, spaces = 2)
+        output.split("\n").each { |l| l.prepend(' ' * (level * spaces)) }.join("\n")
+      end
+    end
+  end
 end