+++ test/cookbooks/test/recipes/default.rb
@@ -1,2 +1,6 @@
-# frozen_string_literal: true
 include_recipe 'unbound::default'
+
+unbound_service 'unbound' do
+  action :enable
+  delayed_action :start
+end